pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "leotraven", email = "leo@traven.de"},
]
dependencies = [
    "langchain>=0.0.341",
    "faiss-cpu>=1.7.4",
    "openai>=1.3.5",
    "sentence-transformers>=2.2.2",
    "llama-cpp-python>=0.2.20",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

